requirements.txt
python-dotenv==1.0.1  # For reading environment variables stored in .env file
langchain==0.2.2
langchain-community==0.2.3
langchain-openai==0.1.8  # For embeddings
unstructured==0.14.4  # Document loading
chromadb==0.5.0  # Vector storage
openai==1.31.1  # For embeddings
tiktoken==0.7.0  # For embeddings
PyPDF2==3.0.1  # PDF reader
google-generativeai==0.3.0  # Google Generative AI for embeddings
streamlit==1.25.0  # Streamlit framework

# install markdown depenendies with: `pip install "unstructured[md]"` after install the requirements file. Leave this line commented out. 
